Refactored Demonic Tyrant handling from DemoPets
